docs: remove scary sounding warnings
The library is now tested in async-hwi/Liana, browsers, and browser
extensions. It is coming out of the experimental stage and is likely
going to work in production as a replacement for the current NPM lib.

The API is unlikely to change significantly, and the semver version is
still at 0.X.Y, which anyway indicates that some breaking changes can
be expected.