restructure README's
The current README for GitHub is not suitable to be the readme at
https://crates.io/crates/bitbox-api or to be the README for the NPM
package (not yet published).

We put the README for crates.io into README-rust.md and reference it
in Cargo.toml.

The one for GitHub goes to .github/README.md, which takes priority on
GitHub.

The one for NPM is in the root README.md, as NPM does not support
specifying another README out of the box.