NPM version: add explicit version configuration
wasm-pack was creating the NPM package from Rust, by adopting the
version from Cargo.toml file. However, both libraries should have their
own version and changelogs. Now, a specific NPM version is written into
NPM_VERSION file, and jq reads this version when `make wasm` is run.
Then this version is added to `package.json` of the build.

The reason to separate the versions and changelogs of Rust and NPM is
that a feature might affect only one of the packages. Then we need a
release in that package, and we do nothing in the other one. Let's say
now a new change in the other package patches a small bug in the system.
It's unclear how the version or release should be updated now. Example:
At the beginning Rust and NPM are on both v1.0.0. After Rust feature
addition, Rust is on v2.0.0 with new release, NPM is on v1.0.0 still.
After NPM bug patch, NPM needs a release but bumping NPM to v2.0.1 is
not a good practice since it did not have a major change. This problem
would not occur if they had separate versions.

Signed-off-by: asi345 <inanata15@gmail.com>