wasm: add a close() method
This closes the connection and calls the onCloseCb callback. The
callback is called manually for webHID (the webHID disconnect callback
is not fired on close), and automatically for the bridge (the socket
onclose event is automatically firex on close).

This allows one to create a short lived connection (release the
BitBox02 after performing the desired operation). This is helpful in
webclients such as Rabby where one only needs to open a quick
connection to e.g. sign a transaction.

This only applies to WASM, not to Rust. In Rust, one can close the
connection by simply dropping the object.